Android SUPL NI client (Revision 1.1)
Network Initiated support for AGPS on UMTS (SUPL) is added to Android framework and the
GPS Hardware Abstraction Layer (HAL). It displays a notification message on the UI when
receiving a notification only NI request, or a dialog box when receiving a notification/verification
request. User responses are sent back to the HAL and GPS libraries (e.g. libloc_api).

Rev 1.1: Repeated closure of NI dialog is prevented.

CRs-Fixed: 186692