Change to dummy wake lock functionality for 8660
Due to the complexities of getting wake locks working in 8660 and the
fact that wake locks are going to be deprecated, there is no motivation
to make them work on 8660 - they are due to be replaced by suspend blockers.
For this reason this is a change to allow wake locks to be avoided in the
HAL layer for 8660.  This change will be removed at such a time as
suspend blockers become available.

Change-Id: Ib924e4b49ccb9a31c16bf57cee436fb7979ff08f