libhardware: Fix Voip Downlink issue
- In Voip call, Downlink path is not
using MVS path.
- In Voip usecase, Downlink path is
not going through MVS and using normal
playback path.
This is due to failure in needsDirectOutput
function. mPhoneState is not updated to
MODE_IN_COMM by the time needsDirectOutput
is called. This condition check is not
required for Voip usecase.
- Remove the condition for mPhoneState in
needsDirectOutput which establishes the Voip
Downlink path via MVS.

CRs-Fixed: 381566
Change-Id: Id65b08977ed3a25b701669b9cbdc768daa1e048c