audio policy: fix in call audio setup delay
There is a delay inserted by audio policy manager before switching
device when entering in call mode to allow music playback to be muted
before the device actually switches.

This delay can be significant when high latency outputs (e.g. deep buffers) are
present and should not be applied if the output is not active.

Issue 7612431.

Change-Id: I04b666769033dbfdc861752b030d3f7bdffe9435