wifi/bt: Add support for BT/WLAN status sync
Since the power consumption issue for BT if the WLAN chip is QCA solution
on the HSIC bus, we need to have some ways to let HSIC bus enter suspend.
Currently, bluetooth stack has to load/unload the WLAN driver when BT is
on and WLAN is off. In order to avoid the sync issue, add the status sync
implementation.

Signed-off-by: Ming-yi Lin <mylin@codeaurora.org>
Change-Id: If54a857d4104cf2d1acb4475ded2eb8c38fe6c10