Use WIFI_DRIVER_STATE_CTRL_PARAM during driver load.
This shall ensure to have a synchronous initialization
for the scenario where the driver is loaded only once
during the boot up.
The wait for completion of the initialization is removed
from the insmod and hence this callback to the driver on
every Turn ON shall ensure a synchronous initialization.
From the driver perspective , the wait for completion of
the initialization is moved from the insmod callback to
this callback (WIFI_DRIVER_STATE_CTRL_PARAM).

CRs-Fixed: 2010050
Change-Id: I33dccc7d7cdca8af6fb9f11175143c5c8eaccc5a