wifi: Temporaty Fix to ensure all the dependencies for LOUN get merged
The changes in wifi.c depend on external projects to get the
WIFI_DRIVER_MODULE_PATH and the init scripts defined to load the driver.
While these changes are getting merged, the Wi-Fi functionality would be
broken. Hence, add a temporary FIX to the existing logic to consider the
scenarios of the above projects not having the changes, thus providing
the flexibility to get all these changes independently.
Please note that this is a temporary FIX to get the things done.

Change-Id: I6f8e6987d466eeec07619fe3e7a97191f3ccbf3d
CRs-Fixed: 2029369